+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.4.6 - 
 * Add support for -SkipCertificateCheck
 * Format RestApi error details rather than the raw json (still want to add a try/catch before release.)
+* `Set-MITFolderAcl`: Add `-FolderPermissions` string[] parameter
+* `New-MITFolderAcl`: Add `-FolderPermissions` string[] parameter
 ## 0.4.5 - Sep 2023
 * `Set-MITFileContent`: Make HashType a parameter
 * `Set-MITFileContent`: Use chunked for files -ge 20MB (previously -ge 2GB)